Methods to the Madness
I added methods to most of the classes except the map class, where I
plan to put all of the graphics stuff.  I also added comments and the
globe constructor.
I also have a lead as to how to do graphics, so I deleted some of the
JFrame and JPanel stuff, because I don’t think I need all of that.