Inheriting some Fancy Graphics
I decided to ignore the broken plate tectonics and spend some time on
the graphics.  I threw in a new projection, polar, and changed the
whole graphics system to use inheritance.  I originally intended Map to
be an interface, but it actually has some methods I would like the
projections to inherit, so I left it as its own class.

P.S. Do not use FromSpace, Hemispherical, Robinson, or Sinusoidal as
they are currently works in progress.