Biomes and Bugfixes!
I fixed the thing where everything was completely broken.
Specifically, I
found and repaired a bug with the smooth(double amount) method (45 min),
found and repaired a bug with the orographic effect (15 min),
moved some code from the Globe class to the main class (15 min),
and rebalanced working smooth and rough methods to create many tiny
islands in the ocean (45 min).
It is not exactly as I would like it right now, but I expect that to
change when I repair the plate tectonics method, which I plan to do
after I get rain() and biomeAssign() down.
I then tried to commit but failed to get a good internet connection, so
this all happened on a later date:
I created and refined a program to create biomes (45 min).
I also made the code more precise and efficient in general (15 min)
and tried but failed to find the problem with plate tectonics (15 min).