Ranges and Trenches and Chains, Oh My!
I did it!  I got terrain generator to work!  Well, kind of.
I ran the terrain generator today and found out that the
plateTechtonics() method had fixed itself and now worked.  It only goes
west, though.  Nothing is affected by what is to its east. (15 min)
I adapted the rest of the code to work with the new plateTechtonics().
(30 min)
I also revised biomeAssign() and refined it to the point where it looks
pretty nice. (60 min)