Gimme a Break
I found the problem with the plate tectonics.  As it turns out, “break”
exits the innermost for, while, or switch statement, not the innermost
brackets as I originally thought.  I rewrote it to use “if” statements
(15 min) and rebalanced some of the other code (30 min).
I also recently found out that I can use “continue” instead of “break”
or “if” statements to solve the problem more elegantly.  I will do that
later.