What is Doing On?
I tried to do some bug fixing and figure out why the plateTechtonics()
are not behaving quite right.  I set up a bunch of System.out.println
statements and figured out that the problem is delTheta, a vector that
should represent the displacement between two tiles and should invert
when looking in the other direction, if that makes any sense.  The
point is, it’s not doing that.  It’s doing something really weird.  At
least I found the source of the problem.  My program is also definitely
noticeably faster. (45 min)