Make it Rain
I got rain() to work!  Kind of.  I have yet to adjust it but it stopped
giving me errors and as far as I can tell, everything is working
perfectly!  I just need to get the rivers to be more predominant over
the lakes. (about 60 min of thinking followed by 30 min of coding)