I Give This Program A C++
I have “finished” rain().  The quotation marks are because I don’t
really like how short and stubby the rivers are, but I have dubbed it
satisfactory, so I am moving on to phase II.  I have a few ideas that
might make it better, but I think I need to move on right now, so I
will return to this later.