Let Us Make Peace
I made wars end.  I finally accomplished this by just disabling the
part that makes wars spread.  This is not a permanent solution, but it
works.  I am thinking of having some kind of system to make the fronts
spread and dissolve on alternating turns to make the fronts more
continuous. (30 min)