I Don't Understand!
How can a program throw an error from inside a try-catch? Isn't the
whole purpose of a try-catch structure to prevent the error from being
thrown?! And how can the color parameters be out of range when all three
of them are 0?! Please explain this to me, because I don't understand at
all!
I tried to start with the galactic generation, which caused me to
realize that lattitude and longitude no longer applied as they did with
a globe. As a result, I had to replace a lot of my methods to use
Points, which should have been simple. After all, I wasn't replacing the
bulk of the code. But then, THIS. I don't even know anymore. Just, ugh!