Down to Earth
I reprogrammed CiviInSpace to be its own program. I could not make it
freestanding without a substantial amount of work, so as it stands,
Civi-HD can be run for the original Earth-based Civilization generator.
In order to see Civi in Space, one must manually replace both instances
of "Globe" in World with "Disc". I may make that runable on its own at a
later date.