That moment you realize there's a problem with a fundamental method that you wrote on the first day of this project.
I created a new river-generation algorithm that creates lakes before
bothering with rivers. In doing so, I discovered a horrible bug: my
adjacentTo() method was wrong. I spent a good long time overhauling many
of the basic methods in Globe.class and finally came to the conclusion
that this method is too slow. On the bright side, my code is much
cleaner and more efficient now.